spawning in gazebo support
The file with inertias and collisions is showing a robot the wheels of
which are not spaced out at the sides of the robot. Need to check.

The error reported is:
[gzserver-1] Warning [Model.cc:216] Non-unique name[left_wheel] detected
2 times in XML children of model with name[box_bot-27682].
[gzserver-1] Warning [Model.cc:216] Non-unique name[right_wheel]
detected 2 times in XML children of model with name[box_bot-27682].
[gzserver-1] Error [Model.cc:317] Joint with name [left_wheel] in model
with name [box_bot-27682] has a name collision. Please rename this
joint.
[gzserver-1] Error [Model.cc:317] Joint with name [right_wheel] in model
with name [box_bot-27682] has a name collision. Please rename this
joint.